refactor: Code improved
The source code for all classes and main file are improved
with some changes.

Option for both async and sync inference are added by using
req_get() method in helper class.

In Helper class, the check_model() method is implemented and is
used in where needed.
Also, the preprocess_input() method is improved and the new method
get_pr_im() is added to support it.